refactor(frontend): extract TemplateCard and form utils; GroupCard with ConfirmDialog
